Jetpack Error Manage PAge: Extracted feature gate example code to parents
